+++ model_builder/domain/reference_data/modeling_templates/other/__init__.py
@@ -1,0 +1,4 @@
+"""Ad-hoc "other" modeling templates (interface-owned, currently empty)."""
+from model_builder.domain.reference_data.modeling_templates.other.registry import OTHER_TEMPLATES
+
+__all__ = ["OTHER_TEMPLATES"]